% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds_ade_ats.R
\name{bounds_ade_ats}
\alias{bounds_ade_ats}
\title{Bounds on direct effect for always-takers}
\usage{
bounds_ade_ats(
  df,
  d,
  m,
  y,
  at_group = 1,
  max_defier_share = 0,
  allow_min_defiers = TRUE,
  reg_formula = NULL,
  num_gridpoints = 10^5
)
}
\arguments{
\item{df}{A data frame}

\item{d}{Name of the treatment variable in the df}

\item{m}{Name of the mediator variable}

\item{y}{Name of the outcome variable}

\item{at_group}{The AT group of interest. Default is 1, so we compute means for units with M(1)=M(0)=1.}

\item{max_defier_share}{Bound on the proportion of defiers in the population. Default is 0 which indicates that the monotonicity constraint is imposed.}

\item{allow_min_defiers}{If the bound on defiers (max_defier_share) is inconsistent with the data, proceed by allowing the minimum number of defiers compatible with the data. Otherwise, throw an error. Default is TRUE.}

\item{reg_formula}{Optional regression formula used to model the joint distribution of (Y, M) given D. If NULL, empirical distributions are used.}

\item{num_gridpoints}{(Optional.) The number of gridpoints used in evaluating the integral. Higher is more accurate but more computationally costly}
}
\description{
Computes bounds on the average direct effect for k-always-takers, \eqn{E[Y(1, k) - Y(1, k) \mid G = kk]}, allowing for optional regression-adjusted probability estimates.
}
